encapsulated logic in ToolTip component
