--- src/App.css
@@ -1,135 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-:root {
-  --color-blue: #4635b1;
-  --color-purple: #b771e5;
-  --color-green: #aeea94;
-  --color-yellow: #fffbca;
-}
-
-main {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-  background: var(--color-purple);
-  height: 100vh;
-  width: 100vw;
-  gap: 5px;
-}
-
-.card {
-  background-color: var(--color-yellow);
-  display: grid;
-  place-items: center;
-}
-
-button {
-  background-color: var(--color-green);
-  color: var(--color-blue);
-}
-
-span {
-  background-color: var(--color-purple);
-  color: var(--color-yellow);
-  padding: 4px 8px;
-  border-radius: 12px;
-  cursor: default;
-}
-
-.maindiv {
-  position: relative;
-}
-.top {
-  color: black;
-  position: absolute;
-  top: -40px;
-  left: 14px;
-  border: 2px solid;
-  background-color: red;
-  border-radius: 5px;
-  z-index: 2;
-}
-
-.right {
-  color: black;
-  position: absolute;
-  top: 6.5px;
-  right: -95px;
-  border: 2px solid;
-  background-color: rgb(238, 163, 163);
-  border-radius: 5px;
-  z-index: 2;
-}
-
-.bottom {
-  color: black;
-  position: absolute;
-  top: 40px;
-  left: 14px;
-  border: 2px solid;
-  background-color: rgb(238, 163, 163);
-  border-radius: 5px;
-  z-index: 2;
-}
-.left {
-  color: black;
-  position: absolute;
-  top: 0px;
-  left: -95px;
-  border: 2px solid;
-  background-color: red;
-  border-radius: 5px;
-  z-index: 2;
-}
-.top p {
-  padding: 2px 2px 2px 2px;
-}
-.sideDraweMainDiv {
-  height: 100vh;
-  width: 100vw;
-  background-color: rgb(0, 0, 0, 0.5);
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  position: relative;
-  z-index: 3;
-}
-
-.sideDrawer {
-  padding: 0;
-  margin: 0;
-  width: 200px;
-  height: 100%;
-  background-color: aquamarine;
-  position: absolute;
-  top: 0px;
-
-  z-index: 3;
-  color: black;
-  animation: slide-in 1s forwards;
-}
-
-@keyframes slide-in {
-  0% {
-    right: -200px;
-  }
-  100% {
-    right: 0px;
-  }
-}
-
-.sidebarContainer {
-  width: 100%;
-  height: 100vh;
-  background-color: rgba(27, 31, 29, 0.5);
-  position: absolute;
-  z-index: 3;
-}
-
-.maincontainer {
-  position: relative;
-}
